[cxx-interop] Fix tests that passed `-module-name Swift`
`Swift` is the name of the Swift standard library module. Creating another module with this name breaks assumptions in the compiler and caused these two tests to fail.
